Fix failing tests. Closes issue 159.
test_formats__audio.TAudioFile.test_performers was failing because
the ~performers and ~people tags were title-cased.

test_browsers__search.TSearchBar was failing because it tried to
access a SongLibrarian; normally that's set up when the main
library is created (library.init()), but since that doesn't happen
in the tests, it must be manually created.