Ensure NumSearch tags compare against a number.
Update issue 216
Labels: Milestone-2.1 -Milestone-2.0
The fix is in for 2.0.1, but the root of this problem lies in the fact that
it is possible to receive different types of objects from song.get() depending
on the song being queried. Fixing this may involve behavior changes, and so
will be in a later version.