Change gtk.main_quit to main.destroy outside of the main window (lazka).
Update issue 223
Status: Fixed
This is fixed to my satisfaction. It's possible to launch QL, queue something,
then quit before the library rescan is complete and not update the on-disk
queue, which should be investigated as it could be a possible conflict between
quitting and the copool that could cause problems.