Implement the playlist model in python using gtk.GenericTreeModel.
CustomModel implements a list model on top of a python list (instead of the internal linked list)
CustomListStore provides a gtk.ListStore API.
The remaining code is split up in a class for tracking a row (current song) and the play order logic.

Speedup for filling is 2x (on an i7), Redrawing/Scrolling should also be faster.
Move/insert/append is slower (python list...).