Support displaying embedded images on Windows. Closes issue 725.
We use NamedTemporaryFile() to have a real file, but on
Windows this can't be reopened, so if loading fails,
load from the open file handle.

Once we have another source for cover URIs (like a http server) we
should try not to write a file at all.. but this works for now.